+++ src/main/java/ortus/boxlang/modules/oracle/OracleDriver.java
@@ -61,12 +61,6 @@ public class OracleDriver extends GenericJDBCDriver {
 
 	@Override
 	public String buildConnectionURL( DatasourceConfig config ) {
-		// Validate the database
-		String database = ( String ) config.properties.getOrDefault( "database", "" );
-		if ( database.isEmpty() ) {
-			throw new IllegalArgumentException( "The database property is required for the Oracle JDBC Driver" );
-		}
-
 		// Validate the host
 		String host = ( String ) config.properties.getOrDefault( "host", DEFAULT_HOST );
 		if ( host.isEmpty() ) {
@@ -95,13 +89,31 @@ public class OracleDriver extends GenericJDBCDriver {
 			);
 		}
 
+		// Validate the database
+		String sid = ( String ) config.properties.getOrDefault( "SID", "" );
+		if ( !sid.isBlank() ) {
+			// Build the connection URL with the SID
+			return String.format(
+			    "jdbc:oracle:%s:@%s:%s:%s",
+			    protocol,
+			    host,
+			    port,
+			    sid
+			);
+		}
+
+		String serviceName = ( String ) config.properties.getOrDefault( "serviceName", "" );
+		if ( serviceName.isBlank() ) {
+			throw new IllegalArgumentException( "Either the serviceName or SID property is required for the Oracle JDBC Driver." );
+		}
+
 		// Build the connection URL
 		return String.format(
 		    "jdbc:oracle:%s:@//%s:%s/%s",
 		    protocol,
 		    host,
 		    port,
-		    database
+		    serviceName
 		);
 	}