+++ src/test/java/ortus/boxlang/modules/oracle/OracleDriverTest.java
@@ -29,22 +29,47 @@ public class OracleDriverTest {
 	}
 
 	@Test
-	@DisplayName( "Test buildConnectionURL()" )
-	public void testBuildConnectionURL() {
+	@DisplayName( "Test connection URLs with SID" )
+	public void testConnectionURLWithSID() {
 		OracleDriver		driver	= new OracleDriver();
 		DatasourceConfig	config	= new DatasourceConfig();
-		config.properties.put( "driver", "oracle" );
-		config.properties.put( "database", "mydb" );
+		config.properties.put( "SID", "mySID" );
+		config.properties.put( "protocol", "thin" );
 
-		String expectedURL = "jdbc:oracle:thin:@//localhost:1521/mydb";
+		String expectedURL = "jdbc:oracle:thin:@localhost:1521:mySID";
 		assertThat( driver.buildConnectionURL( config ) ).isEqualTo( expectedURL );
 	}
 
-	@DisplayName( "Throw an exception if the database is not found" )
 	@Test
-	public void testBuildConnectionURLNoDatabase() {
+	@DisplayName( "Test connection URLs with service name" )
+	public void testConnectionURLWithServiceName() {
 		OracleDriver		driver	= new OracleDriver();
 		DatasourceConfig	config	= new DatasourceConfig();
+		config.properties.put( "serviceName", "myServiceName" );
+		config.properties.put( "protocol", "thin" );
+
+		String expectedURL = "jdbc:oracle:thin:@//localhost:1521/myServiceName";
+		assertThat( driver.buildConnectionURL( config ) ).isEqualTo( expectedURL );
+	}
+
+	@DisplayName( "Throw an exception if the serviceName or SID is not found" )
+	@Test
+	public void testBuildConnectionURLNoServiceNameOrSID() {
+		OracleDriver		driver	= new OracleDriver();
+		DatasourceConfig	config	= new DatasourceConfig();
+
+		assertThrows( IllegalArgumentException.class, () -> {
+			driver.buildConnectionURL( config );
+		} );
+	}
+
+	@DisplayName( "Throw an exception if the serviceName or SID are both blank or empty" )
+	@Test
+	public void testBuildConnectionURLBlankServiceNameOrSID() {
+		OracleDriver		driver	= new OracleDriver();
+		DatasourceConfig	config	= new DatasourceConfig();
+		config.properties.put( "serviceName", "" );
+		config.properties.put( "sid", "   " );
 
 		assertThrows( IllegalArgumentException.class, () -> {
 			driver.buildConnectionURL( config );